requirements.txt
aiohttp==3.8.1
asyncio==3.4.3
discord.py==2.0.1
requests==2.28.1
yarl==1.8.1